added dataset download ability to testing endpoint moved the common logic to query module
